requirements.txt
pynput>=1.7.6
selenium>=4.0.0
webdriver-manager>=4.0.0
openai>=1.0.0
google-generativeai>=0.3.0
pillow>=10.0.0
pyautogui>=0.9.54
playwright>=1.46.0
customtkinter>=5.0.0
SpeechRecognition>=3.10.0
python-dotenv>=1.0.0
python-docx>=1.0.0
python-pptx>=0.6.0
openpyxl>=3.1.0
mss>=9.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
sounddevice>=0.4.6
soundfile>=0.12.1
pyinstaller>=6.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
